@@ -1,50 +1,64 @@
 {
-  "name": "course-project",
-  "version": "0.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "~7.2.0",
-    "@angular/common": "~7.2.0",
-    "@angular/compiler": "~7.2.0",
-    "@angular/core": "~7.2.0",
-    "@angular/forms": "~7.2.0",
-    "@angular/platform-browser": "~7.2.0",
-    "@angular/platform-browser-dynamic": "~7.2.0",
-    "@angular/router": "~7.2.0",
-    "@ngrx/store": "^9.2.0",
-    "bootstrap": "^3.4.1",
-    "core-js": "^2.5.4",
-    "rxjs": "~6.3.3",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "~7.3.10",
-    "@angular/compiler-cli": "~7.2.0",
-    "@angular/language-service": "~7.2.0",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.2.2"
-  }
+    "name": "course-project",
+    "version": "0.0.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e",
+        "compile:server": "webpack --config webpack.server.config.js --progress --colors",
+        "serve:ssr": "node dist/server",
+        "build:ssr": "npm run build:client-and-server-bundles && npm run compile:server",
+        "build:client-and-server-bundles": "ng build --prod && ng run courseProject:server:production"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular-devkit/build-angular": "^0.803.24",
+        "@angular/animations": "^8.0.0",
+        "@angular/common": "^8.0.0",
+        "@angular/compiler": "^8.0.0",
+        "@angular/core": "^8.0.0",
+        "@angular/forms": "^8.0.0",
+        "@angular/platform-browser": "^8.0.0",
+        "@angular/platform-browser-dynamic": "^8.0.0",
+        "@angular/platform-server": "^8.2.14",
+        "@angular/router": "^8.0.0",
+        "@nestjs/ng-universal": "^1.2.0",
+        "@ngrx/effects": "^8.0.0",
+        "@ngrx/router-store": "^8.0.0",
+        "@ngrx/store": "^8.0.0",
+        "@nguniversal/express-engine": "^7.0.0",
+        "@nguniversal/module-map-ngfactory-loader": "^7.1.1",
+        "bootstrap": "^3.4.1",
+        "core-js": "^2.5.4",
+        "express": "^4.17.1",
+        "rxjs": "~6.3.3",
+        "tslib": "^1.9.0",
+        "zone.js": "~0.8.26"
+    },
+    "devDependencies": {
+        "@angular/cli": "^8.0.0",
+        "@angular/compiler-cli": "^8.0.0",
+        "@angular/language-service": "^8.0.0",
+        "@ngrx/store-devtools": "^8.0.0",
+        "@types/jasmine": "~2.8.8",
+        "@types/jasminewd2": "~2.0.3",
+        "@types/node": "~8.9.4",
+        "codelyzer": "~4.5.0",
+        "jasmine-core": "~2.99.1",
+        "jasmine-spec-reporter": "~4.2.1",
+        "karma": "~4.0.0",
+        "karma-chrome-launcher": "~2.2.0",
+        "karma-coverage-istanbul-reporter": "~2.0.1",
+        "karma-jasmine": "~1.1.2",
+        "karma-jasmine-html-reporter": "^0.2.2",
+        "protractor": "~5.4.0",
+        "ts-loader": "^5.2.0",
+        "ts-node": "~7.0.0",
+        "tslint": "~5.11.0",
+        "typescript": "~3.5.3",
+        "webpack-cli": "^3.1.0"
+    }
 }
