@@ -1,16 +1,21 @@
 import { Injectable } from '@angular/core';
 import { HttpInterceptor, HttpHandler, HttpRequest, HttpParams } from '@angular/common/http';
 import { AuthService } from './auth.service';
-import { exhaustMap, take } from 'rxjs/operators';
+import { exhaustMap, take, map } from 'rxjs/operators';
+import { Store } from '@ngrx/store';
+import * as fromApp from '../store/app.reducer';
 
 @Injectable()
 export class AuthInterceptorService implements HttpInterceptor {
 
-    constructor(private authService: AuthService) { }
+    constructor(private authService: AuthService, private store: Store<fromApp.AppState>) { }
 
     intercept(req: HttpRequest<any>, next: HttpHandler) {
-        return this.authService.user.pipe(
+        return this.store.select('auth').pipe(
             take(1),
+            map(authState => {
+                return authState.user;
+            }),
             exhaustMap(user => {
                 if (!user) {
                     return next.handle(req);
