@@ -1,6 +1,9 @@
-import { Component, OnInit } from '@angular/core';
-import { AuthService } from './auth/auth.service';
+import { Component, Inject, OnInit, PLATFORM_ID } from '@angular/core';
+import { isPlatformBrowser } from '@angular/common';
+import { Store } from '@ngrx/store';
+import * as AuthActions from './auth/store/auth.actions';
 import { LoggingService } from './logging.service';
+import * as fromApp from './store/app.reducer';
 
 @Component({
   selector: 'app-root',
@@ -9,12 +12,18 @@ import { LoggingService } from './logging.service';
 })
 export class AppComponent implements OnInit {
 
-  constructor(private authService: AuthService, private loggingService: LoggingService) {
+  constructor(
+    private loggingService: LoggingService,
+    private store: Store<fromApp.AppState>,
+    @Inject(PLATFORM_ID) private platformId
+  ) {
 
   }
 
   ngOnInit() {
-    this.authService.autoLogin();
+    if (isPlatformBrowser(this.platformId)) {
+      this.store.dispatch(new AuthActions.AutoLogin());
+    }
     this.loggingService.printLog('Hello from AppComponent');
   }
 
