@@ -1,147 +1,28 @@
-import { HttpClient, HttpErrorResponse } from '@angular/common/http';
 import { Injectable } from "@angular/core";
-import { Router } from '@angular/router';
-import { BehaviorSubject, throwError } from 'rxjs';
-import { catchError, tap } from 'rxjs/operators';
-import { User } from './user.model';
-import { environment } from '../../environments/environment';
-
-export interface AuthResponseData {
-    kind: string,
-    idToken: string,
-    email: string,
-    refreshToken: string,
-    expiresIn: string,
-    localId: string,
-    registered?: boolean
-}
+import { Store } from '@ngrx/store';
+import * as fromApp from '../store/app.reducer';
+import * as AuthActions from './store/auth.actions';
 
 @Injectable({ providedIn: 'root' })
 export class AuthService {
 
-    user = new BehaviorSubject<User>(null);
-    tokenExpirationTimer: any;
+    private tokenExpirationTimer: any;
 
-    constructor(private http: HttpClient, private router: Router) {
+    constructor(
+        private store: Store<fromApp.AppState>
+    ) { }
 
-    }
-
-    singup(email: string, password: string) {
-        return this.http.post<AuthResponseData>(
-            'https://identitytoolkit.googleapis.com/v1/accounts:signUp?key=' + environment.firebaseAPIKey,
-            {
-                email: email,
-                password: password,
-                returnSecureToken: true
-            }).pipe(
-                catchError(this.handleError),
-                tap(resData => {
-                    this.handleAuthentication(
-                        resData.email,
-                        resData.localId,
-                        resData.idToken,
-                        +resData.expiresIn
-                    );
-                })
-            );
-    }
-
-    login(email: string, password: string) {
-        return this.http
-            .post<AuthResponseData>(
-                'https://identitytoolkit.googleapis.com/v1/accounts:signInWithPassword?key=' + environment.firebaseAPIKey,
-                {
-                    email: email,
-                    password: password,
-                    returnSecureToken: true
-                }
-            )
-            .pipe(
-                catchError(this.handleError),
-                tap(resData => {
-                    this.handleAuthentication(
-                        resData.email,
-                        resData.localId,
-                        resData.idToken,
-                        +resData.expiresIn
-                    );
-                })
-            );
-    }
-
-    autoLogin() {
-        const userData: {
-            email: string;
-            id: string;
-            _token: string;
-            _tokenExpirationDate: string;
-        } = JSON.parse(localStorage.getItem('userData'));
-
-        if (!userData) {
-            return;
-        }
-
-        const loadedUser = new User(
-            userData.email,
-            userData.id,
-            userData._token,
-            new Date(userData._tokenExpirationDate)
-        );
-
-        if (loadedUser.token) {
-            this.user.next(loadedUser);
-            const expirationDuration =
-                new Date(userData._tokenExpirationDate).getTime() -
-                new Date().getTime();
-            this.autoLogout(expirationDuration);
-        }
-    }
-
-    logout() {
-        this.user.next(null);
-        this.router.navigate(['/auth']);
-        localStorage.removeItem('userData');
-        if (this.tokenExpirationTimer) {
-            clearTimeout(this.tokenExpirationTimer);
-        }
-        this.tokenExpirationTimer = null;
-    }
-
-    autoLogout(expirationDuration: number) {
+    setLogoutTimer(expirationDuration: number) {
         this.tokenExpirationTimer = setTimeout(() => {
-            this.logout();
+            this.store.dispatch(new AuthActions.Logout());
         }, expirationDuration);
     }
 
-    private handleAuthentication(
-        email: string,
-        userId: string,
-        token: string,
-        expiresIn: number
-    ) {
-        const expirationDate = new Date(new Date().getTime() + expiresIn * 1000);
-        const user = new User(email, userId, token, expirationDate);
-        this.user.next(user);
-        this.autoLogout(expiresIn * 1000);
-        localStorage.setItem('userData', JSON.stringify(user));
+    clearLogoutTimer() {
+        if (this.tokenExpirationTimer) {
+            clearTimeout(this.tokenExpirationTimer);
+            this.tokenExpirationTimer = null;
+        }
     }
 
-    private handleError(errorRes: HttpErrorResponse) {
-        let errorMessage = 'An unknown error ocurred!';
-        if (!errorRes.error || !errorRes.error.error) {
-            return throwError(errorMessage);
-        }
-        switch (errorRes.error.error.message) {
-            case 'EMAIL_EXISTS':
-                errorMessage = 'This email exists already';
-                break;
-            case 'EMAIL_NOT_FOUND':
-                errorMessage = 'The email does not exist.';
-                break;
-            case 'INVALID_PASSWORD':
-                errorMessage = 'The password is not correct.';
-                break;
-        }
-        return throwError(errorMessage);
-    }
 }
