@@ -1,7 +1,10 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
-import { DataStorageService } from '../shared/data-storage.service';
-import { AuthService } from '../auth/auth.service';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { Store } from '@ngrx/store';
 import { Subscription } from 'rxjs';
+import { map } from 'rxjs/operators';
+import * as AuthActions from '../auth/store/auth.actions';
+import * as RecipeActions from '../recipes/store/recipe.actions';
+import * as fromApp from '../store/app.reducer';
 
 @Component({
     selector: 'app-header',
@@ -12,16 +15,21 @@ export class HeaderComponent implements OnInit, OnDestroy {
     isAuthenticated = false;
     private userSub: Subscription;
 
-    constructor(private dataStorageService: DataStorageService, private authService: AuthService) {
+    constructor(
+        private store: Store<fromApp.AppState>
+    ) {
 
     }
 
     ngOnInit() {
-        this.userSub = this.authService.user.subscribe(user => {
-            this.isAuthenticated = !!user;
-            console.log(!user);
-            console.log(!!user);
-        });
+        this.userSub = this.store
+            .select('auth')
+            .pipe(map(authState => authState.user))
+            .subscribe(user => {
+                this.isAuthenticated = !!user;
+                console.log(!user);
+                console.log(!!user);
+            });
     }
 
     ngOnDestroy() {
@@ -29,14 +37,16 @@ export class HeaderComponent implements OnInit, OnDestroy {
     }
 
     onSaveData() {
-        this.dataStorageService.storeRecipes();
+        //this.dataStorageService.storeRecipes();
+        this.store.dispatch(new RecipeActions.StoreRecipes());
     }
 
     onFetchData() {
-        this.dataStorageService.fetchRecipes().subscribe();
+        //this.dataStorageService.fetchRecipes().subscribe();
+        this.store.dispatch(new RecipeActions.FetchRecipes());
     }
 
-    onLogout(){
-        this.authService.logout();
+    onLogout() {
+        this.store.dispatch(new AuthActions.Logout());
     }
 }
