@@ -1,23 +1,39 @@
 import { Injectable } from '@angular/core';
+import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
+import { Actions, ofType } from '@ngrx/effects';
+import { Store } from '@ngrx/store';
+import { take, map, switchMap } from 'rxjs/operators';
+import * as RecipesActions from '../recipes/store/recipe.actions';
+import * as fromApp from '../store/app.reducer';
 import { Recipe } from './recipe.model';
-import { Resolve, RouterStateSnapshot, ActivatedRouteSnapshot } from '@angular/router';
-import { Observable } from 'rxjs';
-import { DataStorageService } from '../shared/data-storage.service';
-import { RecipeService } from './recipe.service';
+import { of } from 'rxjs';
 
 @Injectable({ providedIn: 'root' })
 export class RecipesResolverService implements Resolve<Recipe[]>{
 
-    constructor(private dataStorageService: DataStorageService, private recipeService: RecipeService) { }
+    constructor(
+        private store: Store<fromApp.AppState>,
+        private actions$: Actions) { }
 
-    resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Recipe[] | Observable<Recipe[]> | Promise<Recipe[]> {
+    resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot) {
 
-        const recipes = this.recipeService.getRecipes();
-
-        if (recipes.length === 0) {
-            return this.dataStorageService.fetchRecipes();
-        } else {
-            return recipes;
-        }
+        //return this.dataStorageService.fetchRecipes();
+        return this.store.select('recipes').pipe(
+            take(1),
+            map(recipesState => {
+                return recipesState.recipes;
+            }),
+            switchMap(recipes => {
+                if (recipes.length === 0) {
+                    this.store.dispatch(new RecipesActions.FetchRecipes());
+                    return this.actions$.pipe(
+                        ofType(RecipesActions.SET_RECIPES),
+                        take(1)
+                    );
+                } else {
+                    return of(recipes);
+                }
+            })
+        );
     }
 }
