@@ -1,14 +1,18 @@
 import { ActionReducerMap } from '@ngrx/store';
 import * as fromAuth from '../auth/store/auth.reducer';
 import * as fromShoppingList from '../shopping-list/store/shopping-list.reducer';
+import * as fromRecipes from '../recipes/store/recipe.reducer';
 
 
 export interface AppState {
     shoppingList: fromShoppingList.State;
     auth: fromAuth.State;
+    recipes: fromRecipes.State;
+
 }
 
 export const appReducer: ActionReducerMap<AppState> = {
     shoppingList: fromShoppingList.shoppingListReducer,
-    auth: fromAuth.authReducer
+    auth: fromAuth.authReducer,
+    recipes: fromRecipes.recipeReducer
 };
